+++ src/backend/activation_codes/tests/test_models.py
@@ -1,5 +1,6 @@
 """Tests for activation_codes models."""
 
+import json
 from datetime import timedelta
 
 from django.core.exceptions import ValidationError
@@ -7,12 +8,18 @@ from django.db.models import ProtectedError
 from django.utils import timezone
 
 import pytest
+import responses
 
 from core.factories import UserFactory
 
 from activation_codes.exceptions import InvalidCodeError, UserAlreadyActivatedError
 from activation_codes.factories import ActivationCodeFactory, UserActivationFactory
-from activation_codes.models import ActivationCode, UserActivation, generate_activation_code
+from activation_codes.models import (
+    ActivationCode,
+    UserActivation,
+    UserRegistrationRequest,
+    generate_activation_code,
+)
 
 
 @pytest.mark.django_db
@@ -270,3 +277,40 @@ def test_user_activation_ordering():
 
     activations = list(UserActivation.objects.all())
     assert activations == [activation2, activation1]
+
+
+@responses.activate
+@pytest.mark.django_db(transaction=True)
+def test_activation_code_use_success_notify_brevo(settings):
+    """Test successfully using an activation code and notify Brevo."""
+    settings.BREVO_API_KEY = "test_brevo_api_key"
+    settings.BREVO_WAITING_LIST_ID = "test_waiting_list_id"
+    settings.BREVO_FOLLOWUP_LIST_ID = "test_followup_list_name"
+
+    brevo_remove_mock = responses.post(
+        "https://api.brevo.com/v3/contacts/lists/test_waiting_list_id/contacts/remove",
+        json={"message": "Contacts added successfully"},
+        status=201,
+    )
+
+    brevo_add_mock = responses.post(
+        "https://api.brevo.com/v3/contacts/lists/test_followup_list_name/contacts/add",
+        json={"message": "Contacts added successfully"},
+        status=201,
+    )
+
+    user = UserFactory()
+    registration = UserRegistrationRequest.objects.create(user=user)
+    activation_code = ActivationCodeFactory()
+    activation = activation_code.use(user)
+
+    registration.refresh_from_db()
+    assert registration.user_activation == activation
+
+    assert len(brevo_remove_mock.calls) == 1
+    assert brevo_remove_mock.calls[0].request.headers["api-key"] == "test_brevo_api_key"
+    assert json.loads(brevo_remove_mock.calls[0].request.body) == {"emails": [user.email]}
+
+    assert len(brevo_add_mock.calls) == 1
+    assert brevo_add_mock.calls[0].request.headers["api-key"] == "test_brevo_api_key"
+    assert json.loads(brevo_add_mock.calls[0].request.body) == {"emails": [user.email]}