+++ src/backend/core/tests/authentication/test_backends.py
@@ -1,5 +1,6 @@
 """Unit tests for the Authentication Backends."""
 
+import json
 import random
 import re
 
@@ -491,3 +492,70 @@ def test_authentication_session_tokens(django_assert_num_queries, monkeypatch, r
     assert user is not None
     assert request.session["oidc_access_token"] == "test-access-token"
     assert get_oidc_refresh_token(request.session) == "test-refresh-token"
+
+
+@responses.activate
+def test_authentication_user_added_to_brevo(monkeypatch, rf, settings):
+    """
+    Test that a user is added to the Brevo follow-up list upon authentication.
+    """
+    settings.OIDC_OP_TOKEN_ENDPOINT = "http://oidc.endpoint.test/token"
+    settings.OIDC_OP_USER_ENDPOINT = "http://oidc.endpoint.test/userinfo"
+    settings.OIDC_OP_JWKS_ENDPOINT = "http://oidc.endpoint.test/jwks"
+
+    settings.BREVO_API_KEY = "test-api-key"
+    settings.BREVO_FOLLOWUP_LIST_ID = "follow-up-list-id"
+    settings.ACTIVATION_REQUIRED = False
+
+    brevo_mock = responses.post(
+        "https://api.brevo.com/v3/contacts/lists/follow-up-list-id/contacts/add",
+        status=400,
+    )
+
+    klass = OIDCAuthenticationBackend()
+    request = rf.get("/some-url", {"state": "test-state", "code": "test-code"})
+    request.session = {}
+
+    def verify_token_mocked(*args, **kwargs):
+        return {"sub": "123", "email": "test@example.com"}
+
+    monkeypatch.setattr(OIDCAuthenticationBackend, "verify_token", verify_token_mocked)
+
+    responses.add(
+        responses.POST,
+        re.compile(settings.OIDC_OP_TOKEN_ENDPOINT),
+        json={
+            "access_token": "test-access-token",
+            "refresh_token": "test-refresh-token",
+        },
+        status=200,
+    )
+
+    responses.add(
+        responses.GET,
+        re.compile(settings.OIDC_OP_USER_ENDPOINT),
+        json={"sub": "123", "email": "test@example.com"},
+        status=200,
+    )
+
+    user = klass.authenticate(
+        request,
+        code="test-code",
+        nonce="test-nonce",
+        code_verifier="test-code-verifier",
+    )
+
+    assert len(brevo_mock.calls) == 1
+    assert brevo_mock.calls[0].request.headers["api-key"] == "test-api-key"
+    assert json.loads(brevo_mock.calls[0].request.body) == {"emails": [user.email]}
+
+    # Now test when activation is required: user should not be added to Brevo list
+    settings.ACTIVATION_REQUIRED = True
+    klass.authenticate(
+        request,
+        code="test-code",
+        nonce="test-nonce",
+        code_verifier="test-code-verifier",
+    )
+
+    assert len(brevo_mock.calls) == 1  # No new call made